juno/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.1
django-cors-headers==3.4.0
django-cors-middleware==1.5.0
django-filter==2.3.0
django-heroku==0.3.1
django-oauth-toolkit==1.3.2
django-polymorphic==1.3
django-rest-coreapi-schema==0.1.2
djangorestframework==3.7.7
drfdocs==0.0.11
heroku==0.1.4
idna==2.10
install==1.3.3
itypes==1.2.0
Jinja2==2.11.2
Markdown==2.6.10
MarkupSafe==1.1.1
oauthlib==3.1.0
psycopg2==2.8.5
Pygments==2.2.0
python-dateutil==1.5
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
whitenoise==5.2.0
